dictionary: Convenient programmatic construction
Variadic constructors have been added to dictionary to facilitate
convenient construction in code, including within a #codeStream entry.
The constructors take an even number of arguments, alternating between
the key and the corresponding value. The values may, themselves, be
dictionaries constructed in the same way. This means that the code
directly maps to the resulting nested dictionary structure.

For example, the following code stream entry:

    #codeStream
    {
        code
        #{
            writeEntry
            (
                os,
                "dict",
                dictionary
                (
                    "s", 1,
                    "wl", wordList({"apples", "oranges"}),
                    "subDict", dictionary
                    (
                        "v", vector(2, 3, 4),
                        "t", tensor(5, 6, 7, 8, 9, 10, 11, 12, 13),
                        "ll", labelList(10, -1)
                    )
                )
            );
        #};
    }

Expands to the following:

    dict
    {
        s               1;
        wl              2 ( apples oranges );
        subDict
        {
            v               ( 2 3 4 );
            t               ( 5 6 7 8 9 10 11 12 13 );
            ll              10 { -1 };
        }
    }