solver: Added optional run-time setting of deltaTFactor
The deltaTFactor used in the automatic time-step adjustment to smoothly increase
the time-step when permitted can now be specified in the system/controlDict but
defaults to 1.2 as used previously.  For example in the
tutorials/incompressibleVoF/damBreak/damBreak case the rate at which the
time-step is increased after the slamming event can be reduced to 10% per
time-step by setting

.
.
.
maxDeltaT       1;
deltaTFactor    1.1;

Alternatively the case can be set to continue with the smallest time-step
without further adjustment by setting deltaTFactor to 1.