intersectionPatchToPatch: Prevent substantial error in face centres
A new algorithm for calculating a face's area and centre has been
implemented as a static method of the face class, called
face::areaAndCentreStabilised. This method tracks the round off error
associated with the calculation of the centroid and uses this to
determine when and how much to stabilise the result.

This is utilised in the calculation of the centroids of face-face
intersections, which become the face centres on non-conformal
couples. This therefore improves the robustness of non-conformal
couples.

This new algorithm has not replaced the procedure used in the standard
calculation of face centres. The face centres in the polygonal mesh are
expected to be of better quality than those resulting from an
intersection calculation, so the additional expense associated with the
stabilised algorithm is not thought to be necessary.