CloudFunctionObjects/VolumeFraction: Replacement for VoidFraction
The volumeFraction cloud function will create and write a vol-field of
the cloud's volume fraction. It requires no additional controls. Example
in constant/cloudProperties:

    cloudFunctions
    {
        volumeFraction1
        {
            type volumeFraction;
        }
    }

This replaces voidFraction, which was incorrectly named in that it also
write the volume fraction (not one-minus the volume fraction), and who's
storage of the field was flawed in that it caused registration clashes
with fields stored by certain other Lagrangian models.