fieldMapper: Removed hasUnmapped
This is not used anywhere. This logic can and should be kept internal to
the mapper. Eventually what will be needed is additional interface to
allow calling code to provide appropriate values to set on unmapped
elements, if there are any. The calling code, however, should not need
to know if any such unmapped elements are actually present.